Add config templates
A config template is a config that is used as a template that settings
from the current actual task config are written over. They can be used
to share site parse settings for the HTML source or any other shared
state. They are found in templates/ dir inside $XDG_CONFIG_HOME