Don't duplicate error message
if an error is critical, its message will already be printed on exit by
eyre. There's no reason to tracing::error!() it, too.

Before the process didn't shutdown if a task encountered a critical
error, thus there could've been merit in just printing that error
instead. Now the process exits as soon as there is a critical error in
any running task and thus there's no reason to print the error message
manually